Fix a warning on clang. Cast to size_t explicitly.

+++ src/glow/Backends/Interpreter/InterpreterNodes.cpp
@@ -866,9 +866,8 @@ void Interpreter::fwdLocalResponseNormalizationInst(
         for (size_t c = 0; c < idim.c; c++) {
           float squareSum = 0.0;
           for (int i = (c >= halfWindowSize ? c - halfWindowSize : 0);
-               i <= std::min(c + halfWindowSize, idim.c - 1);
-               i++) { 
-            auto val = inW.at({n, h, w, i});
+               i <= std::min(c + halfWindowSize, idim.c - 1); i++) {
+            auto val = inW.at({n, h, w, (size_t)i});
             squareSum += val * val;
           }
